requirements.txt
flask
gunicorn
python-dotenv
google-genai
requests